cv_analysis_backend/requirements.txt
Flask==2.3.2
pandas==1.5.3
python-dotenv==1.0.0
PyPDF2==3.0.1
PyMuPDF
python-docx==0.8.11
psycopg2-binary==2.9.6
Werkzeug==2.3.3
scikit-learn==1.4.0
flask-cors==4.0.1
PyJWT
mongoengine
pymongo
Flask_login
cloudmersive-virus-api-client==3.2.0
requests==2.31.0
gunicorn==20.1.0
numpy==1.25.1
sentence-transformers
spacy
tensorflow-hub
pdfplumber==0.11.4
torch
torchvision
nltk

https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0.tar.gz
